+++ uefi-raw/CHANGELOG.md
@@ -17,6 +17,9 @@
 - Added  `DevicePathProtocol::length()` properly constructing the `u16` value
 
 ## Changed
+- **Breaking:** Types `Ipv4Address` and `Ipv6Address` have been removed. They
+  were replaced by `core::net::Ipv4Addr` and `core::net::Ipv6Addr`, as they are
+  ABI compatible. This mainly affects the `IpAddress` wrapper type.
 - `DevicePathProtocol` now derives
   `Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash`